feat: support to set params of remoting client
